pyproject.toml
[tool.poetry]
name = "cloudfspy"
version = "0.1.0"
description = ""
authors = ["haruiz <henryruiz22@gmail.com>"]
readme = "README.md"
packages = [{ include = "cloudfspy", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
google-cloud-storage = { version = "^2.14.0", optional = true }
azure-storage-blob = { version = "^12.19.0", optional = true }
boto3 = { version = "^1.34.35", optional = true }
rich = "^13.7.0"
tqdm = "^4.66.1"
python-dotenv = "^1.0.1"
colorama = "^0.4.6"


[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
mkdocs-material = "^9.5.11"
mkdocs-glightbox = "^0.3.7"
mkdocs-video = "^1.5.0"
jupyterlab = "^4.1.2"
mkdocs-jupyter = "^0.24.6"
notebook = "^7.1.0"

[tool.poetry.extras]
gs = ["google-cloud-storage"]
s3 = ["boto3"]
azure = ["azure-storage-blob"]
all = ["google-cloud-storage", "boto3", "azure-storage-blob"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
